move jsx components to component folder
