+++ src/App.js
@@ -1,47 +1,36 @@
-import React, { useState } from "react";
-//import "./App.css";
-import LessonsData from "./LessonsData.json";
-import Header from "./component/header/Header";
+import React from "react";
+import Navbar from "./component/header/Navbar";
 import Home from "./component/home/Home";
+import LessonsData from "./LessonsData.json";
 import Quiz from "./component/quiz/Quiz";
 import Lessons from "./component/lessons/Lessons";
 import LessonDetail from "./component/lessons/LessonDetail";
+import Post from "./component/home/Post";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
+import "./app.css";
 
-function App() {
-  const [name, setName] = useState("");
-  const [difficulty, setDifficulty] = useState("");
-
+const App = () => {
   return (
     <BrowserRouter>
-      <Header />
-      <Routes>
-        <Route
-          path="/"
-          element={
-            <Home
-              name={name}
-              setName={setName}
-              setDifficulty={setDifficulty}
-              difficulty={difficulty}
-            />
-          }
-        />
-        <Route
-          path="/questions/lessons/:id"
-          element={<Quiz name={name} difficulty={difficulty} />}
-        />
-        <Route
-          path="/lessons"
-          element={<Lessons LessonsData={LessonsData} />}
-        />
-        <Route
-          path="/lessons/:id"
-          element={<LessonDetail LessonsData={LessonsData} />}
-        />
-      </Routes>
+      <div>
+        <Navbar user="" />
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route
+            path="/lessons"
+            element={<Lessons LessonsData={LessonsData} />}
+          />
+          <Route
+            path="/lessons/:id"
+            element={<LessonDetail LessonsData={LessonsData} />}
+          />
+          <Route path="/questions/lessons/:id" element={<Quiz />} />
+
+          <Route path="/post/:id" element={<Post />} />
+        </Routes>
+      </div>
     </BrowserRouter>
   );
-}
+};
 
 export default App;
