--- src/component/header/Header.css
@@ -1,3 +1,0 @@
-.header {
-  text-align: center;
-}
