+++ src/component/home/Home.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Card from "../components/Card";
-import { posts } from "../data";
+import Card from "./Card";
+import { posts } from "../../data";
 
 const Home = () => {
   return (
