+++ src/component/home/Post.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { useLocation } from "react-router";
-import { posts } from "../data";
+import { posts } from "../../data";
 
 const Post = () => {
   const location = useLocation();
