+++ src/index.js
@@ -2,7 +2,7 @@
 // import { createRoot } from 'react-dom/client';
 // import ReactDOM from "react-dom/client";
 import ReactDOM from "react-dom";
-import App from "./App.jsx";
+import App from "./App";
 
 import { Auth0Provider } from "@auth0/auth0-react";
